Run fsck in CopyEfsTest
We're copying from a live r/w image, so the picture we get may be a bit
distorted. Running fsck over the image first should help with
reliability, but f2fs does some fixes on mount as well, so we mount and
unmount the image before running dump.f2fs.

Change-Id: Ibb15388c55b6f909a1d1b849f9760f248644d9e9
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 349120101
Test: atest CopyEfsTest
Flag: TEST_ONLY